MobiusTransformations

Documentation for MobiusTransformations.

MobiusTransformations.MöbiusTransformation

MobiusTransformations.MöbiusTransformation

Base.:*

Base.:∘

Base.isone

LinearAlgebra.normalize

MobiusTransformations.Möbius

MobiusTransformations.Möbius

MobiusTransformations.Möbius

MobiusTransformations.Möbius

MobiusTransformations.Möbius

MobiusTransformations.Möbius

MobiusTransformations.Möbius

MobiusTransformations.set_infinity

MobiusTransformations.MöbiusTransformation — Type

MöbiusTransformation{T}

Represents a Möbius transformation of the form: f(z) = (az + b) / (cz + d)

Fields

a::T, b::T, c::T, d::T: Coefficients of the transformation

MobiusTransformations.MöbiusTransformation — Method

(m::MöbiusTransformation)(z)

Apply to a number z the Möbius transformation m(z) = (a*z + b) / (c*z + d), where m = Möbius([a b; c d]).

Base.:* — Method

*(m::MöbiusTransformation, n::MöbiusTransformation)

Compose two Möbius transformations.

Base.:∘ — Method

∘(m::MöbiusTransformation, n::MöbiusTransformation)

Compose two Möbius transformations (same as *).

Base.isone — Method

isone(m::MöbiusTransformation)

Return true if m is the identity Mobius transformation and false otherwise.

LinearAlgebra.normalize — Method

normalize(m::MöbiusTransformation)

Returns a Möbius transformation m2 such that m2 == m and det(m2) = 1. Requires sqrt(det(m)) to be defined in the coefficient field.

MobiusTransformations.Möbius — Method

Möbius(a, b, c, d)

Creates a MöbiusTransformation with coefficients a, b, c, d.

MobiusTransformations.Möbius — Method

Möbius(x, y, z, X, Y, Z)

Returns the Möbius transformation that maps [x, y, z] to [X, Y, Z]. Values of Inf are permitted.

MobiusTransformations.Möbius — Method

Möbius(M::AbstractMatrix)

Creates a Möbius transformation with coefficients a, b, c, d, where M=[a b; c d].

MobiusTransformations.Möbius — Method

Möbius(x, y, z)

Returns the Möbius transformation that maps [0, 1, Inf] to points [x, y, z]. Values of Inf are permitted.

MobiusTransformations.Möbius — Method

Möbius(source, target)

Returns the Möbius transformation that maps source to target. Values of Inf are permitted.

MobiusTransformations.Möbius — Method

Möbius(target)

Returns the Möbius transformation that maps [0, 1, Inf] to target = [x, y z]. Values of Inf are permitted.

MobiusTransformations.Möbius — Method

Möbius(::Type{T}=ComplexF64)

Returns the identity Möbius transformation of type T.

MobiusTransformations.set_infinity — Method

set_infinity(infinity)

Sets the representation of infinity used by the package.